Move testing utils to be able to use them in different modules
