+++ ktor-core/test/org/jetbrains/ktor/tests/application/ApplicationRequestHeaderTest.kt
@@ -3,7 +3,7 @@ package org.jetbrains.ktor.tests.application
 import org.jetbrains.ktor.application.*
 import org.jetbrains.ktor.http.*
 import org.jetbrains.ktor.routing.*
-import org.jetbrains.ktor.tests.*
+import org.jetbrains.ktor.testing.*
 import org.junit.*
 import kotlin.test.*
