+++ ktor-core/test/org/jetbrains/ktor/tests/application/HandlerTest.kt
@@ -2,7 +2,7 @@ package org.jetbrains.ktor.tests.application
 
 import org.jetbrains.ktor.application.*
 import org.jetbrains.ktor.http.*
-import org.jetbrains.ktor.tests.*
+import org.jetbrains.ktor.testing.*
 import org.junit.*
 import kotlin.test.*
