+++ ktor-core/test/org/jetbrains/ktor/tests/http/ContentTypeTest.kt
@@ -1,7 +1,7 @@
 package org.jetbrains.ktor.tests.http
 
 import org.jetbrains.ktor.http.*
-import org.jetbrains.ktor.tests.*
+import org.jetbrains.ktor.testing.*
 import org.junit.*
 import kotlin.test.*
